* changed filename due to udev update (Bug #16571)
git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/trunk/ucs@13938 1b283449-9f2f-0410-b571-8eb228e1a901

+++ base/univention-installer/scripts/83_filter_udev_rules.sh
@@ -32,10 +32,10 @@
 
 UDEVDIR="/instmnt/etc/udev/rules.d"
 mkdir -p "$UDEVDIR"
-export UDEVRULEFN="${UDEVDIR}/z25_persistent-net.rules"
+export UDEVRULEFN="${UDEVDIR}/70-persistent-net.rules"
 
 # if dummy network interface is in use, delete mapping from existing rules file
-if [ -f "/tmp/dummy-network-interface.txt" ] ; then
+if [ -f "/tmp/dummy-network-interface.txt" -a -f "$UDEVRULEFN" ] ; then
 	MACADDR=$(/bin/ifconfig eth0 | grep " HWaddr " | awk "{ print $NF }")
 	TMPFN=$(mktemp /tmp/temp.XXXXXXX)
 	cat $UDEVRULEFN | grep -v "eth0" | grep -v "$MACADDR" > $TMPFN
